+++ script_robot_example01.py
@@ -1,8 +1,17 @@
-from R12 import RobotBat2
+import numpy
 
+from R12 import RobotBat, Target
+import time
 from matplotlib import pyplot
 
-R = RobotBat2.RobotBat()
-R.initialize()
-R.set_position(1000, 100, 300, 0, 0)
-data = R.measure(plot=True)
+
+R = RobotBat.RobotBat()
+#R.initialize()
+
+for target_x in numpy.linspace(-500,500, 10):
+
+    target = Target.Target(x=target_x, y=700, z = 350, rotation=-90)
+    result = target.get_robot_position(distance=500,  azimuth=0)
+    R.set_position(result)
+    time.sleep(1)
+    data = R.measure(plot=True)
